This is a version that works as I understand it; however EUnit fails.
I think this is probably because tests are run in parallel, while processes
are registered with a name, so I should probably disable parallel testing.

This will be something for tomorrow, as the Mrs. has been nagging me on and
of for spending some time on the PC today ;-).

Nevertheless it was quite a fun exercise todo.